chore: cleanup, split out source code from hpp
