+++ components/edgelab/src/algorithm/el_algorithm_base.hpp
@@ -27,9 +27,7 @@
 #define _EL_ALGORITHM_BASE_HPP_
 
 #include <cstdint>
-#include <forward_list>
 
-#include "el_debug.h"
 #include "el_inference_base.hpp"
 #include "el_types.h"
 
@@ -85,74 +83,6 @@ class Algorithm {
     uint32_t __run_time;          // ms
     uint32_t __postprocess_time;  // ms
 };
-
-Algorithm::Algorithm(EngineType* engine, const InfoType& info)
-    : __p_engine(engine),
-      __p_input(nullptr),
-      __algorithm_info(info),
-      __preprocess_time(0),
-      __run_time(0),
-      __postprocess_time(0) {
-    __input_shape  = engine->get_input_shape(0);
-    __output_shape = engine->get_output_shape(0);
-    __input_quant  = engine->get_input_quant_param(0);
-    __output_quant = engine->get_output_quant_param(0);
-}
-
-Algorithm::~Algorithm() {
-    __p_engine = nullptr;
-    __p_input  = nullptr;
-
-    __preprocess_time  = 0;
-    __run_time         = 0;
-    __postprocess_time = 0;
-}
-
-el_err_code_t Algorithm::underlying_run(void* input) {
-    el_err_code_t ret{EL_OK};
-    uint32_t      start_time{0};
-    uint32_t      end_time{0};
-
-    EL_ASSERT(__p_engine != nullptr);
-
-    __p_input = input;
-
-    // preprocess
-    start_time        = el_get_time_ms();
-    ret               = preprocess();
-    end_time          = el_get_time_ms();
-    __preprocess_time = end_time - start_time;
-
-    if (ret != EL_OK) {
-        return ret;
-    }
-
-    // run
-    start_time = el_get_time_ms();
-    ret        = __p_engine->run();
-    end_time   = el_get_time_ms();
-    __run_time = end_time - start_time;
-
-    if (ret != EL_OK) {
-        return ret;
-    }
-
-    // postprocess
-    start_time         = el_get_time_ms();
-    ret                = postprocess();
-    end_time           = el_get_time_ms();
-    __postprocess_time = end_time - start_time;
-
-    return ret;
-}
-
-Algorithm::InfoType Algorithm::get_algorithm_info() const { return __algorithm_info; };
-
-uint32_t Algorithm::get_preprocess_time() const { return __preprocess_time; }
-
-uint32_t Algorithm::get_run_time() const { return __run_time; }
-
-uint32_t Algorithm::get_postprocess_time() const { return __postprocess_time; }
 
 }  // namespace base
