+++ components/edgelab/src/algorithm/el_algorithm_pfld.hpp
@@ -27,12 +27,9 @@
 #define _EL_ALGORITHM_PFLD_HPP_
 
 #include <cstdint>
-#include <type_traits>
-#include <utility>
+#include <forward_list>
 
 #include "el_algorithm_base.hpp"
-#include "el_cv.h"
-#include "el_debug.h"
 #include "el_types.h"
 
 namespace edgelab::algorithm {
@@ -53,7 +50,7 @@ class PFLD : public edgelab::algorithm::base::Algorithm {
     using PointType  = el_point_t;
     using ConfigType = types::el_algorithm_pfld_config_t;
 
-    static constexpr InfoType algorithm_info{types::el_algorithm_pfld_config_t::info};
+    static  InfoType algorithm_info;
 
     PFLD(EngineType* engine);
     PFLD(EngineType* engine, const ConfigType& config);
@@ -64,7 +61,7 @@ class PFLD : public edgelab::algorithm::base::Algorithm {
     el_err_code_t                       run(ImageType* input);
     const std::forward_list<PointType>& get_results() const;
 
-    void       set_algorithm_config(const ConfigType& config);
+    void       set_algorithm_config(const ConfigType&);
     ConfigType get_algorithm_config() const;
 
    protected:
@@ -80,109 +77,6 @@ class PFLD : public edgelab::algorithm::base::Algorithm {
 
     std::forward_list<PointType> _results;
 };
-
-PFLD::PFLD(EngineType* engine)
-    : edgelab::algorithm::base::Algorithm(engine, PFLD::algorithm_info), _w_scale(1.f), _h_scale(1.f) {
-    EL_ASSERT(is_model_valid(engine));
-    init();
-}
-
-PFLD::PFLD(EngineType* engine, const ConfigType& config)
-    : edgelab::algorithm::base::Algorithm(engine, config.info), _w_scale(1.f), _h_scale(1.f) {
-    EL_ASSERT(is_model_valid(engine));
-    init();
-}
-
-PFLD::~PFLD() { _results.clear(); }
-
-bool PFLD::is_model_valid(const EngineType* engine) {
-    const auto& input_shape{engine->get_input_shape(0)};
-    if (input_shape.size != 4 ||      // B, W, H, C
-        input_shape.dims[0] != 1 ||   // B = 1
-        input_shape.dims[1] < 32 ||   // W >= 32
-        input_shape.dims[2] < 32 ||   // H >= 32
-        (input_shape.dims[3] != 3 &&  // C = RGB or Gray
-         input_shape.dims[3] != 1))
-        return false;
-
-    const auto& output_shape{engine->get_output_shape(0)};
-    if (output_shape.size != 2 ||     // B, PTS
-        output_shape.dims[0] != 1 ||  // B = 1
-        output_shape.dims[1] % 2      // PTS should be multiply of 2
-    )
-        return false;
-
-    return true;
-}
-
-inline void PFLD::init() {
-    _input_img.data   = static_cast<decltype(ImageType::data)>(this->__p_engine->get_input(0));
-    _input_img.width  = static_cast<decltype(ImageType::width)>(this->__input_shape.dims[1]),
-    _input_img.height = static_cast<decltype(ImageType::height)>(this->__input_shape.dims[2]),
-    _input_img.size =
-      static_cast<decltype(ImageType::size)>(_input_img.width * _input_img.height * this->__input_shape.dims[3]);
-    _input_img.format = EL_PIXEL_FORMAT_UNKNOWN;
-    _input_img.rotate = EL_PIXEL_ROTATE_0;
-
-    if (this->__input_shape.dims[3] == 3) {
-        _input_img.format = EL_PIXEL_FORMAT_RGB888;
-    } else if (this->__input_shape.dims[3] == 1) {
-        _input_img.format = EL_PIXEL_FORMAT_GRAYSCALE;
-    }
-
-    EL_ASSERT(_input_img.format != EL_PIXEL_FORMAT_UNKNOWN);
-    EL_ASSERT(_input_img.rotate != EL_PIXEL_ROTATE_UNKNOWN);
-}
-
-el_err_code_t PFLD::run(ImageType* input) {
-    _w_scale = static_cast<float>(input->width) / static_cast<float>(_input_img.width);
-    _h_scale = static_cast<float>(input->height) / static_cast<float>(_input_img.height);
-
-    // TODO: image type conversion before underlying_run, because underlying_run doing a type erasure
-    return underlying_run(input);
-};
-
-el_err_code_t PFLD::preprocess() {
-    auto* i_img{static_cast<ImageType*>(this->__p_input)};
-
-    // convert image
-    rgb_to_rgb(i_img, &_input_img);
-
-    auto size{_input_img.size};
-    for (decltype(ImageType::size) i{0}; i < size; ++i) {
-        _input_img.data[i] -= 128;
-    }
-
-    return EL_OK;
-}
-
-el_err_code_t PFLD::postprocess() {
-    _results.clear();
-
-    // get output
-    auto*   data{static_cast<int8_t*>(this->__p_engine->get_output(0))};
-    float   scale{this->__output_quant.scale};
-    bool    rescale{scale < 0.1f ? true : false};
-    int32_t zero_point{this->__output_quant.zero_point};
-    auto    pred_l{this->__output_shape.dims[1]};
-
-    scale = rescale ? scale * 100.f : scale;
-
-    for (decltype(pred_l) i{0}; i < pred_l; i += 2) {
-        _results.emplace_front(
-          PointType{.x      = static_cast<decltype(PointType::x)>(((data[i] - zero_point) * scale) * _w_scale),
-                    .y      = static_cast<decltype(PointType::y)>(((data[i + 1] - zero_point) * scale) * _h_scale),
-                    .target = static_cast<decltype(PointType::target)>(i / 2)});
-    }
-
-    return EL_OK;
-}
-
-const std::forward_list<PFLD::PointType>& PFLD::get_results() const { return _results; }
-
-void PFLD::set_algorithm_config(const ConfigType& config [[maybe_unused]]) {}
-
-PFLD::ConfigType PFLD::get_algorithm_config() const { return ConfigType{}; }
 
 }  // namespace edgelab::algorithm
