+++ examples/at_server_demo/main/at_utility.hpp
@@ -31,100 +31,6 @@ std::string string_2_str(const std::string& str) {
     return ss;
 }
 
-const char* err_code_2_str(el_err_code_t ec) {
-    switch (ec) {
-    case EL_OK:
-        return "\"ok\"";
-    case EL_AGAIN:
-        return "\"try again\"";
-    case EL_ELOG:
-        return "\"logic error\"";
-    case EL_ETIMOUT:
-        return "\"timeout\"";
-    case EL_EIO:
-        return "\"input/output error\"";
-    case EL_EINVAL:
-        return "\"invalid argument\"";
-    case EL_ENOMEM:
-        return "\"out of memory\"";
-    case EL_EBUSY:
-        return "\"busy\"";
-    case EL_ENOTSUP:
-        return "\"not supported\"";
-    case EL_EPERM:
-        return "\"operation not permitted\"";
-    default:
-        return "\"unknown\"";
-    }
-}
-
-const char* algo_type_2_str(el_algorithm_type_t at) {
-    switch (at) {
-    case EL_ALGO_TYPE_FOMO:
-        return "\"FOMO\"";
-    case EL_ALGO_TYPE_PFLD:
-        return "\"PFLD\"";
-    case EL_ALGO_TYPE_YOLO:
-        return "\"YOLO\"";
-    case EL_ALGO_TYPE_IMCLS:
-        return "\"IMCLS\"";
-    default:
-        return "\"undefined\"";
-    }
-}
-
-const char* algo_cat_2_str(el_algorithm_cat_t ac) {
-    switch (ac) {
-    case EL_ALGO_CAT_DET:
-        return "\"detection\"";
-    case EL_ALGO_CAT_POSE:
-        return "\"pose\"";
-    case EL_ALGO_CAT_CLS:
-        return "\"classification\"";
-    default:
-        return "\"undefined\"";
-    }
-}
-
-const char* sensor_type_2_str(el_sensor_type_t st) {
-    switch (st) {
-    case EL_SENSOR_TYPE_CAM:
-        return "\"camera\"";
-    default:
-        return "\"undefined\"";
-    }
-}
-
-const char* sensor_sta_2_str(el_sensor_state_t ss) {
-    switch (ss) {
-    case EL_SENSOR_STA_REG:
-        return "\"registered\"";
-    case EL_SENSOR_STA_AVAIL:
-        return "\"available\"";
-    case EL_SENSOR_STA_LOCKED:
-        return "\"locked\"";
-    default:
-        return "\"unknown\"";
-    }
-}
-
-const char* img_type_2_str(el_pixel_format_t pix_fmt) {
-    switch (pix_fmt) {
-    case EL_PIXEL_FORMAT_GRAYSCALE:
-        return "\"grayscale\"";
-    case EL_PIXEL_FORMAT_JPEG:
-        return "\"jpeg\"";
-    case EL_PIXEL_FORMAT_RGB565:
-        return "\"rgb565\"";
-    case EL_PIXEL_FORMAT_RGB888:
-        return "\"rgb888\"";
-    case EL_PIXEL_FORMAT_YUV422:
-        return "\"yuv422\"";
-    default:
-        return "\"undefined\"";
-    }
-}
-
 inline uint32_t color_literal(uint8_t i) {
     static const uint16_t color[] = {
       0x0000,
